+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- Include parameters in stack frames. (#662) @Tyrrrz
 - Remove CultureUIInfo if value is even with CultureInfo. (#671) @lucas-zimerman
 - Make all fields on UserFeedback optional. (#660) @Tyrrrz
 - Align transaction names with Java. (#659) @Tyrrrz